+++ packages/lib/src/model/Model.ts
@@ -77,6 +77,8 @@ export function ExtendedModel<TProps extends ModelProps, TBaseModelClass>(
   baseModel: TBaseModelClass,
   modelProps: TProps
 ): _Model<TBaseModelClass & Object extends ModelClass<infer M> ? M : never, TProps> {
+  assertIsModelClass(baseModel, "baseModel")
+
   // note that & Object is there to support abstract classes
   return internalModel<TProps, any>(modelProps, baseModel as any)
 }